# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Mayor votings",
        "title": "Mayor API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/electionPeriod": {
            "post": {
                "description": "Endpoint to insert election periods, should only be used to insert missing/hisotical data",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Election periods"
                ],
                "summary": "Inserts election periods",
                "parameters": [
                    {
                        "description": "the election periods that are going to be inserted",
                        "name": "periods",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ElectionPeriod"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ElectionPeriod"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/electionPeriod/range": {
            "get": {
                "description": "Returns all election periods that took place in a given timespan",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Election periods"
                ],
                "summary": "Get election periods by timespan",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "from The beginning of the selected timespan",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "The end of the selected timespan",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ElectionPeriod"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/electionPeriod/{year}": {
            "get": {
                "description": "Returns the election periods that took place in a given year",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Election periods"
                ],
                "summary": "Get the election period of a certain year",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the searched year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ElectionPeriod"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/mayor/current": {
            "get": {
                "description": "Returns the current mayor",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mayor"
                ],
                "summary": "Get the current mayor",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Candidate"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/mayor/names": {
            "get": {
                "description": "Returns all mayor names",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mayor"
                ],
                "summary": "Get names of all mayors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/mayor/next": {
            "get": {
                "description": "Returns the mayor with the most votes in the current election. If there is currently no election, this returns null.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mayor"
                ],
                "summary": "Get the next mayor",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Candidate"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Candidate": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "perks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Perk"
                    }
                }
            }
        },
        "model.ElectionPeriod": {
            "type": "object",
            "properties": {
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Candidate"
                    }
                },
                "end": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "winner": {
                    "$ref": "#/definitions/model.Candidate"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "model.Perk": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}